fix: filter condtions etc 🐛
