+++ server/graphql/resolvers/index.ts
@@ -34,7 +34,7 @@ export { OutlookCategoryResolver } from './outlookCategory'
 export { ProjectResolver } from './project'
 export { ReportLinkResolver } from './reportLink'
 export { ReportsResolver } from './reports/ReportsResolver'
-export { RoleResolver } from './role'
+export { RoleResolver } from './role/RoleResolver'
 export { SubscriptionResolver } from './subscription'
 export { TimesheetResolver } from './timesheet'
 export * from './types'
